+++ JeonPanGeun/JeonPanGeun/1주차/main.swift
@@ -7,14 +7,4 @@
 
 import Foundation
 
-let n = Int(readLine()!)!
-
-for i in 0..<n {
-    for _ in (0..<i).reversed() {
-        print(" ", terminator: "")
-    }
-    for j in 0..<(n-i) {
-        print("*", terminator: "")
-    }
-    print("")
-}
+print("HI")
